Moving files for this story into data tools
Previously classes were in data/standardisers. They do not perform standardisation functions so are moved to the outer folder
